refactor(platform/os): 拆分 os.rs 为 macos/windows 子模块并加固平台路径
按 oracle 设计将 1755 行的 os.rs 拆分为 7 个子模块，同时修复多项 Windows/macOS 平台问题。

拆分结构（M7）：
- os/mod.rs：facade + fallback impl
- os/macos/mod.rs：所有 macOS 逻辑
- os/windows/mod.rs：进程控制、launch/quit、计划任务等横切 helper
- os/windows/install.rs：安装、exe hash patch、版本记录
- os/windows/restore.rs：恢复、dry-run、错误回滚、目录同步
- os/windows/backup.rs：备份定位/写入（解 install ↔ restore 循环依赖）
- os/windows/permissions.rs：WindowsApps 权限、takeown/icacls

平台修复：
- M3：takeown/icacls 不再吞错（原 let _ = run_command 改为 ?）
- M5：exe hash patch 加固——marker 唯一性校验、原 64 字节 hex 校验、
  使用 Win32 MoveFileExW + MOVEFILE_REPLACE_EXISTING 实现真正原子替换
- S4：windowsapps_permission_targets 用 canonicalize 处理大小写/短路径/符号链接
- S4 后续修复：剥离 \\\\?\\\\ 扩展长度路径前缀再传给 takeown/icacls
  （不剥离会报 File or Directory not found）
- S6：macOS 备份恢复改为取最新备份（与 Windows 一致），
  清理旧备份逻辑修正（原 skip(1) 错误地保留了最旧、清理了刚恢复的最新）

依赖：
- crates/platform/Cargo.toml 新增 windows crate 的 Win32_Storage_FileSystem
  feature 用于 MoveFileExW

测试：
- 新增 patch_exe_hash_data 4 个单测（空/多 marker、非 hex、正常替换）
- 新增 strip_unc_prefix 3 个单测
- 既有 sync_dir_exact / cleanup / restore_from_snapshot 测试归位到 restore.rs
- windowsapps_permissions_include_app_dir_for_exe_rewrite 标记 ignore
  （依赖真实 WindowsApps 目录，仅生产环境运行）

验证：
- cargo check -p claude-zh-platform 通过
- cargo test -p claude-zh-platform 通过（30 passed / 1 ignored）
- cargo clippy --workspace -- -D warnings 通过
