+++ crates/platform/Cargo.toml
@@ -22,6 +22,7 @@ windows = { version = "0.58", features = [
   "Win32_Foundation",
   "Win32_Globalization",
   "Win32_Security",
+  "Win32_Storage_FileSystem",
   "Win32_System_Registry",
   "Win32_System_Threading",
   "Win32_UI_Shell"
